+++ code/elife/news-server/src/main/java/newsserver/controller/NoticeController.java
@@ -33,39 +33,47 @@ public class NoticeController {
 
     @RequestMapping(path = "/addNotice")
     @ResponseBody
-    public JSONObject addNotice(@RequestParam String content, @RequestParam String managerName, @RequestParam int communityId, @RequestParam String username, @RequestParam int isMass){
-        net.minidev.json.JSONObject object = new net.minidev.json.JSONObject();
-        if("".equals(content)||"".equals(managerName)||communityId==0){
-            object.put("addNotice", "0");
-            object.put("message","信息不能为空");
+    public JSONObject addNotice(HttpServletRequest request, @RequestParam String content, @RequestParam String managerName, @RequestParam int communityId, @RequestParam String username, @RequestParam int isMass){
+        HttpSession session = request.getSession();
+        String name = (String) session.getAttribute("username");
+        String role = (String) session.getAttribute("role");
+        if (StringUtils.isEmpty(name) || !("1".equals(role)) || !name.equals(managerName)) {
+            JSONObject jsonObject = new JSONObject();
+            jsonObject.put("login", 0);
+            return jsonObject;
+        } else {
+            net.minidev.json.JSONObject object = new net.minidev.json.JSONObject();
+            if ("".equals(content) || "".equals(managerName) || communityId == 0) {
+                object.put("addNotice", "0");
+                object.put("message", "信息不能为空");
+                return object;
+            }
+            SimpleDateFormat df = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+            String noticeTime = df.format(new Date());
+            Notice notice = new Notice(noticeTime, content, managerName, communityId);
+            /**先在notice数据库中存*/
+            int noticeId = noticeService.save(notice);
+
+            if (isMass == 0) {
+                NoticeUser noticeUser = new NoticeUser(noticeId, username);
+                noticeService.saveNoticeUser(noticeUser);
+
+            } else {
+                List<String> usernameList = noticeService.findUsernameByCommunityId(communityId);
+                int length = usernameList.size();
+                for (int i = 0; i < length; i++) {
+                    NoticeUser noticeUser1 = new NoticeUser(noticeId, usernameList.get(i));
+                    noticeService.saveNoticeUser(noticeUser1);
+                }
+            }
+            object.put("addNotice", "1");
+            object.put("message", "发布物业通知成功");
             return object;
         }
-        SimpleDateFormat df = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-        String noticeTime = df.format(new Date());
-        Notice notice=new Notice(noticeTime,content,managerName,communityId);
-        /**先在notice数据库中存*/
-         int noticeId = noticeService.save(notice);
-
-         if(isMass==0){
-             NoticeUser noticeUser=new NoticeUser(noticeId,username);
-             noticeService.saveNoticeUser(noticeUser);
-
-         }else {
-             List<String> usernameList=noticeService.findUsernameByCommunityId(communityId);
-             int length=usernameList.size();
-             for (int i=0;i<length;i++){
-                 NoticeUser noticeUser1=new NoticeUser(noticeId,usernameList.get(i));
-                 noticeService.saveNoticeUser(noticeUser1);
-             }
-         }
-        object.put("addNotice", "1");
-        object.put("message","发布物业通知成功");
-        return object;
     }
     @RequestMapping(path = "/ManagerFindNotice")
     @ResponseBody
     public JSONArray  managerFindNotice(HttpServletRequest request, @RequestParam String managerName, @RequestParam int pageNumber, @RequestParam int pageSize){
-
         HttpSession session = request.getSession();
         String name = (String) session.getAttribute("username");
         String role = (String) session.getAttribute("role");
@@ -133,17 +141,25 @@ public class NoticeController {
      * 管理员功能，删除数据库中一条物业信息，包括notice表中和noticeUser表中*/
     @RequestMapping(path = "/deleteOneNotice")
     @ResponseBody
-    public JSONObject deleteOneNotice(@RequestParam int noticeId){
-        net.minidev.json.JSONObject object = new net.minidev.json.JSONObject();
-        if(noticeService.deleteByNotcieId(noticeId).equals("删除这一条物业信息成功")){
-            object.put("deleteOneNotice", "1");
-            object.put("message","删除这一条物业信息成功");
+    public JSONObject deleteOneNotice(HttpServletRequest request, @RequestParam int noticeId){
+        HttpSession session = request.getSession();
+        String name = (String) session.getAttribute("username");
+        String role = (String) session.getAttribute("role");
+        if (StringUtils.isEmpty(name) || !("1".equals(role))) {
+            JSONObject jsonObject = new JSONObject();
+            jsonObject.put("login", 0);
+            return jsonObject;
+        } else {
+            net.minidev.json.JSONObject object = new net.minidev.json.JSONObject();
+            if (noticeService.deleteByNotcieId(noticeId).equals("删除这一条物业信息成功")) {
+                object.put("deleteOneNotice", "1");
+                object.put("message", "删除这一条物业信息成功");
+                return object;
+            }
+            object.put("deleteOneNotice", "0");
+            object.put("message", "删除这一条物业信息失败");
             return object;
         }
-        object.put("deleteOneNotice", "0");
-        object.put("message","删除这一条物业信息失败");
-        return object;
-
     }
     /**
      * 用户功能，删除我的通知列表中的某一条通知*/
